requirements.txt
streamlit==0.72.0
Keras==2.4.3
numpy==1.19.2
tensorflow-cpu==2.3.1
beautifulsoup4==4.9.3
h5py~=2.10.0
lxml==4.6.2